Correction of the output type of the dB_fixed64 function

+++ openair1/PHY/TOOLS/dB_routines.c
@@ -162,7 +162,7 @@ int8_t dB_fixed(uint32_t x)
   return dB_power;
 }
 
-int8_t dB_fixed64(uint64_t x)
+int dB_fixed64(uint64_t x)
 {
   if ((x < (((uint64_t)1) << 32)))
     return (dB_fixed((uint32_t)x));
